src/applications/document_rag/requirements.txt
asyncio==3.4.3
azure-identity==1.23.0
azure-search-documents==11.5.2
azure-storage-blob==12.25.1
langchain==0.3.25
langchain-community==0.3.24
langchain-openai==0.3.16
langgraph==0.4.3
pandas==2.2.3
pydantic-settings==2.9.1
streamlit==1.45.1